+++ CHANGELOG.md
@@ -25,6 +25,8 @@ v4.4.0
 - New options for load balancing (DOF or BOUNDARY) in mesh partitioner (!617)
 - Rework nodal utilities to support nodal prismatic elements (!660)
 - Update Body/Field forces at each timestep (!665)
+- Update nodalutil to include quad and hex elements and introduce SPI nodal
+  points (!696)
 
 **ADRSolver:**
 - Add a projection equation system for C^0 projections (!675)
@@ -41,16 +43,21 @@ v4.4.0
 
 **FieldConvert:**
 - Allow equi-spaced output for 1D and 2DH1D fields (!613)
+- Update quality metric to include scaled Jacobian output (!695)
 
 **NekMesh:**
 - Modify curve module to allow for spline input (!628)
 - Add STL surface writer module (!668)
 - New module for inserting an alternate high-order surface into the working
   mesh (!669)
+- Add curve projection routines to CAD system (!697)
 - Improvements to mesh linearisation module (!659)
 - Add support for Gmsh high-order output (!679)
 - Move CAD classes to factory format (!676)
+- Add module to check topology of the mesh along with boundary connectivity
+  to detect problems such as hanging nodes (!691)
 - Add option to `linearise` module to linearise only prisms (!688)
+- Add option to `linearise` to use element quality (!690)
 
 **FieldConvert:**
 - Move all modules to a new library, FieldUtils, to support post-processing